+++ src/features/Forum/components/Sidebar/Sidebar.stories.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { object, text } from '@storybook/addon-knobs';
-import { topics as topicsMock } from '@features/Forum/mocks';
+import { object } from '@storybook/addon-knobs';
+import { topicList as topicListMock } from '@features/Forum/mocks';
 import { Sidebar } from './Sidebar';
 
 storiesOf('Forum', module).add('Sidebar', () => {
-  const userName = text('userName', 'User');
-  const topics = object('topics', topicsMock);
+  const topicList = object('topicList', topicListMock);
 
   return (
-    <Sidebar topics={topics} userName={userName} />
+    <Sidebar topicList={topicList} />
   );
 });
