--- src/features/Forum/components/Topic/index.ts
@@ -1,2 +1,0 @@
-export * from './Topic';
-export * from './types';
