+++ src/features/Forum/components/TopicItem/TopicItem.css
@@ -1,27 +1,27 @@
-.topic {
+.topic-item {
   border-bottom: 1px solid var(--color-gray);
   cursor: pointer;
   list-style-type: none;
   padding: 1rem 0.6rem;
 }
 
-.topic_selected {
+.topic-item_selected {
   background: var(--color-primary);
 }
 
-.topic__header {
+.topic-item__header {
   align-items: flex-end;
   display: flex;
   justify-content: space-between;
   margin-bottom: 0.25em;
 }
 
-.topic__content {
+.topic-item__content {
   flex: 1;
   min-width: 0;
 }
 
-.topic__name {
+.topic-item__name {
   font-size: var(--font-size-l);
   margin-right: 0.5rem;
   font-weight: bold;
@@ -30,7 +30,7 @@
   white-space: nowrap;
 }
 
-.topic__message {
+.topic-item__message {
   display: block;
   font-size: var(--font-size-m);
   overflow: hidden;
@@ -38,6 +38,6 @@
   white-space: nowrap;
 }
 
-.topic__time {
+.topic-item__time {
   font-size: var(--font-size-m);
 }
