+++ src/features/Forum/components/TopicItem/index.ts
@@ -1,0 +1,2 @@
+export * from './TopicItem';
+export * from './types';
