+++ src/features/Forum/mocks/topicList.ts
@@ -1,4 +1,4 @@
-export const topics = [
+export const topicList = [
   {
     id: 1,
     text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
